Node.js + Express + HBS
We create a node.js solution, which we will be focused on creating a backend Node.js with express in order to respond as a server. This information in mostly returning JSON, HTML or text to front end.

Then we changed all the information to be loaded from HTML, then from HBS, giving the dynamic template some information to be loaded with.

Then we created partials and included them in the hbs documents.

The we created the response for any page that isn't in the app.get, so we could return a 404 error